@@ -1,39 +1,84 @@
 name: flt_imo
 description: A new Flutter project.
 
-publish_to: 'none'
+# The following line prevents the package from being accidentally published to
+# pub.dev using `pub publish`. This is preferred for private packages.
+publish_to: 'none' # Remove this line if you wish to publish to pub.dev
+
+# The following defines the version and build number for your application.
+# A version number is three numbers separated by dots, like 1.2.43
+# followed by an optional build number separated by a +.
+# Both the version and the builder number may be overridden in flutter
+# build by specifying --build-name and --build-number, respectively.
+# In Android, build-name is used as versionName while build-number used as versionCode.
+# Read more about Android versioning at https://developer.android.com/studio/publish/versioning
+# In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
+# Read more about iOS versioning at
+# https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  cupertino_icons: ^1.0.0
-  flutter_speed_dial: null
-  get: ^3.24.0
-  http: null
-  lottie: ^0.7.0+1
+  adaptive_action_sheet: ^2.0.0
+  country_code_picker: ^2.0.1
+  cupertino_icons: ^1.0.2
+  curved_navigation_bar: ^0.4.0-nullsafety.0
+  flutter_speed_dial: ^3.0.5
+  get: ^4.1.4
+  google_mobile_ads: ^0.13.0
   mime: ^1.0.0
-  pin_code_fields: ^3.0.1
-  responsive_framework: ^0.0.14
-  shared_preferences: ^0.5.12+4
-  velocity_x: ^1.0.1
+  pin_code_fields: ^7.1.0
+  responsive_framework: ^0.1.0
+  shared_preferences: ^2.0.5
+  velocity_x: ^3.1.0
+
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  adaptive_action_sheet: null
-  carousel_slider: ^3.0.0
-  country_code_picker: ^1.7.0
-  geocoder: null
-  google_maps_flutter: null
-  google_maps_place_picker: null
-  image_cropper: ^1.3.1
-  image_picker: null
+  geocode: ^1.0.1
+  google_maps_flutter: ^2.0.5
+  google_maps_place_picker: ^2.0.0-nullsafety.3
+  http: ^0.13.3
+  image_cropper: ^1.4.0
+  image_picker: ^0.7.5+2
+  share: ^2.0.4
 
+# For information on the generic Dart part of this file, see the
+# following page: https://dart.dev/tools/pub/pubspec
+# The following section is specific to Flutter.
 flutter:
 
+  # The following line ensures that the Material Icons font is
+  # included with your application, so that you can use the icons in
+  # the material Icons class.
   uses-material-design: true
+  # To add assets to your application, add an assets section, like this:
   assets:
     - assets/images/
+  # An image asset can refer to one or more resolution-specific "variants", see
+  # https://flutter.dev/assets-and-images/#resolution-aware.
+  # For details regarding adding assets from package dependencies, see
+  # https://flutter.dev/assets-and-images/#from-packages
+  # To add custom fonts to your application, add a fonts section here,
+  # in this "flutter" section. Each entry in this list should have a
+  # "family" key with the font family name, and a "fonts" key with a
+  # list giving the asset and other descriptors for the font. For
+  # example:
+  # fonts:
+  #   - family: Schyler
+  #     fonts:
+  #       - asset: fonts/Schyler-Regular.ttf
+  #       - asset: fonts/Schyler-Italic.ttf
+  #         style: italic
+  #   - family: Trajan Pro
+  #     fonts:
+  #       - asset: fonts/TrajanPro.ttf
+  #       - asset: fonts/TrajanPro_Bold.ttf
+  #         weight: 700
+  #
+  # For details regarding fonts from package dependencies,
+  # see https://flutter.dev/custom-fonts/#from-packages
