@@ -1,12 +1,15 @@
+import 'package:flt_imo/Controller/TabViewController/dashboardController.dart';
+import 'package:flt_imo/Models/BoxModel.dart';
 import 'package:flt_imo/Models/inventoryModel.dart';
 import 'package:flt_imo/Models/locationModel.dart';
 import 'package:flt_imo/Service/itemService.dart';
+import 'package:flt_imo/Utils/app_constants.dart';
+import 'package:flt_imo/Utils/mySnackbar.dart';
 import 'package:image_picker/image_picker.dart';
 import 'package:flt_imo/Utils/strings.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import 'package:flt_imo/Utils/mySnackbar.dart';
 import 'package:adaptive_action_sheet/adaptive_action_sheet.dart';
 
 class AddItemController extends GetxController {
@@ -18,6 +21,7 @@ class AddItemController extends GetxController {
   List<dynamic> imageFile = List<dynamic>.empty(growable: true).obs;
   var selectedLocation = Location(name: "Select Location").obs;
   var selectedInventory = Inventory(name: "Select Inventory").obs;
+  var selectedBox = Box(title: "Select Box").obs;
 
   RxInt count = 0.obs;
   TextEditingController titleController = TextEditingController();
@@ -25,35 +29,47 @@ class AddItemController extends GetxController {
   TextEditingController priceController = TextEditingController();
   TextEditingController quantityController = TextEditingController();
 
+  clearParam() {
+    selectedBox.value = Box(title: "Select Box");
+    selectedInventory.value = Inventory(name: "Select Inventory");
+    isSaleable.value = false;
+    isNegotiable.value = false;
+    isFragile.value = false;
+    titleController.text = "";
+    descriptionController.text = "";
+    priceController.text = "";
+    count.value = 0;
+  }
+
   setBody() {
     return ({
-      "BoxKey": "${descriptionController.text}",
+      "BoxKey": "${selectedBox.value.key}",
+      "InventoryId": "${selectedInventory.value.id}",
       "IsSaleable": "${isSaleable.value}",
-      "Price": "${priceController.text}",
+      "ProjectId": "${AppConstants.PROJECT.id}",
+      "Title": "${titleController.text}",
       "IsNegotiable": "${isNegotiable.value}",
+      "Description": "${descriptionController.text}",
       "IsFragile": "${isFragile.value}",
-      "Quantity": "${quantityController.text}",
+      "Price": "${priceController.text}",
+      "Quantity": count.value,
     });
   }
 
   addItem(getAble) {
-    var body;
-    addItemApi(body).then((response) {
-      if (response == null) {
-        processLoading(false);
-        mySnackbar(title: txtFailed, description: txtUnkownError);
-      }
-      if (response.statusCode == 201) {
+    processLoading(true);
+    var body = setBody();
+    ItemService.addItemApi(body).then((response) {
+      if (response != null) {
         snackBarBack(title: txtSuccess, description: "Item Added Successfully").then((value) {
           if (getAble) {
             // Get.find<InventoryListController>().getInventoryList(AppConstants.LOCATION_OBJECT.id.toString());
             Get.back();
+            Get.find<DashboardController>().getRefreshData();
           }
-          processLoading(false);
+          // processLoading(false);
         });
-      } else {
         processLoading(false);
-        mySnackbar(title: txtFailed, description: txtUnkownError);
       }
     }).catchError((e) {
       printError();
@@ -98,7 +114,7 @@ class AddItemController extends GetxController {
 
   Future getImageFromCamera(BuildContext context) async {
     // ignore: deprecated_member_use
-    var image = await ImagePicker.pickImage(source: ImageSource.camera);
+    var image = await ImagePicker().getImage(source: ImageSource.camera);
     Navigator.pop(context);
     if (image != null) {
       imageFile.add(image);
@@ -107,7 +123,7 @@ class AddItemController extends GetxController {
 
   Future getImageFromGallery(BuildContext context) async {
     // ignore: deprecated_member_use
-    var image = await ImagePicker.pickImage(source: ImageSource.gallery);
+    var image = await ImagePicker().getImage(source: ImageSource.gallery);
     Navigator.pop(context);
 
     if (image != null) {
