@@ -1,10 +1,12 @@
+import 'dart:convert';
+
 import 'package:flt_imo/Models/inventoryModel.dart';
 import 'package:flt_imo/Service/inventoryService.dart';
+import 'package:flt_imo/Service/locationService.dart';
 import 'package:flt_imo/Utils/mySnackbar.dart';
 import 'package:flt_imo/Utils/strings.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import 'package:velocity_x/velocity_x.dart';
 
 class InventoryListController extends GetxController {
   var isLoading = true.obs;
@@ -12,15 +14,30 @@ class InventoryListController extends GetxController {
   List<Inventory> inventoryListForDisplay = List<Inventory>.empty(growable: true).obs;
   @override
   void onInit() {
-    getInventoryList();
     super.onInit();
   }
 
   getInventoryList() async {
     isLoading(true);
-    await allInventoryApi().then((response) {
-      if (response.statusCode == 200) {
-        var i = inventoryFromJson(response.body);
+    await InventoryService.allInventoryApi().then((response) {
+      if (response != null) {
+        var i = inventoryFromJson(jsonEncode(response.body));
+        inventoryList = i;
+        print("Inventory : $inventoryList");
+        inventoryListForDisplay = inventoryList;
+      } else {
+        inventoryList = [];
+        inventoryListForDisplay = inventoryList;
+      }
+    });
+    isLoading(false);
+  }
+
+  getInventories(id) async {
+    isLoading(true);
+    await LocationService.locationDetailApi(id).then((response) {
+      if (response != null) {
+        var i = inventoryFromJson(jsonEncode(response.body["inventories"]));
         inventoryList = i;
         print("Inventory : $inventoryList");
         inventoryListForDisplay = inventoryList;
@@ -37,16 +54,10 @@ class InventoryListController extends GetxController {
   }
 
   deleteInventory(BuildContext context, id) async {
-    Future.delayed(Duration(seconds: 1), VxToast.showLoading(context, msg: "Deleting"));
-    deleteInventoryApi(id).then((response) {
-      if (response == null) {
-        mySnackbar(title: txtFailed, description: "Unkown Error Occured");
-      }
-      if (response.statusCode == 200) {
-        mySnackbar(title: txtSuccess, description: "Inventory Deleted Successfully");
+    InventoryService.deleteInventoryApi(id).then((response) {
+      if (response != null) {
+        mySnackbar(title: (txtSuccess), description: "Inventory Deleted Successfully");
         getInventoryList();
-      } else {
-        mySnackbar(title: txtFailed, description: "Unkown Error Occured");
       }
     });
   }
